upgrade to ruff 0.16 and fix the newly-enabled rules (#162)
ruff 0.16 turns on a broad set of rules by default; the repo's config
selects nothing explicitly, so all of them landed at once — 112
violations across 23 rules on a tree that was clean under 0.15.13.

Most were mechanical (import sorting, `int(round(x))` -> `round(x)`,
`datetime.timezone.utc` -> `datetime.UTC`, `zip(xs, xs[1:])` ->
`itertools.pairwise`). Two groups needed judgment:

- SIM115 (23x): every one was `json.load(open(path))`, leaking the
  handle to the GC. Rewritten to the repo's dominant
  `json.loads(path.read_text())` idiom.
- B023 (6x): closures over loop variables. All six are consumed inside
  their own iteration, so none was a live bug, but the bindings are now
  explicit defaults so a later edit cannot silently break them.

Two violations are suppressed with reasons rather than "fixed": the
blind except in _load_page_color_data is a deliberate best-effort load,
and clip_masks' MultiPolygon ValueError is load-bearing — make_iiif_georef
catches ValueError to skip the page, so TRY004's TypeError would turn a
skipped page into a failed run.

710 tests pass, pyright clean.

+++ mapsnap/keymap/crnn_model.py
@@ -268,7 +268,7 @@ def locate_number(
     sx_hi = (last_t + 1) * cell
 
     col0 = max(0, int(sx_lo))
-    col1 = min(width, int(round(sx_hi)))
+    col1 = min(width, round(sx_hi))
     sub = strip[:, col0:col1] if col1 > col0 else strip
     _, ink = cv2.threshold(sub, 0, 255, cv2.THRESH_BINARY_INV | cv2.THRESH_OTSU)
     center = ink_row_center(ink.sum(axis=1) / 255)
